+++ JWT/config/index.js
@@ -1,0 +1,4 @@
+require("dotenv").config();
+
+const jwtSecretKey = process.env.JWT_SECRET_KEY;
+exports.jwtSecretKey = jwtSecretKey;